+++ R/ensureRappDirectoryComponents.r
@@ -6,6 +6,9 @@
 #'   	
 #' @param ns \strong{Signature argument}.
 #'    Object containing namespace information.
+#' @param overwrite \code{\link{logical}}.
+#'    \code{TRUE}: overwrite existing directory components;
+#'    \code{FALSE}: do not overwrite existing directory components.
 #' @template threedot
 #' @example inst/examples/ensureRappDirectoryComponents.r
 #' @seealso \code{
@@ -20,6 +23,7 @@ setGeneric(
   ),
   def = function(
     ns,
+    overwrite = FALSE,
     ...
   ) {
     standardGeneric("ensureRappDirectoryComponents")       
@@ -50,21 +54,12 @@ setMethod(
   ), 
   definition = function(
     ns,
+    overwrite,
     ...
   ) {
     
   if (isPackageProject()) {
     ## Private function //
-    
-    .formatOptionFile <- function(x, name) {
-      expr <- unlist(strsplit(deparse(x, control = NULL), split = ","))
-      expr <- gsub("^\\s*$|(^\\s*(?=\\w))", "", expr, perl = TRUE)
-      expr <- expr[(expr != "")]
-      expr <- unlist(strsplit(expr, split = "(?<=list)\\(|\\)$", perl = TRUE))
-      expr <- paste0(name, " <- ", expr[1], 
-         "(\n\t", paste(expr[2:length(expr)], collapse=",\n\t"), "\n)")
-      expr
-    }
     
     dirs <- c(
       "rapp/.internal",
@@ -75,53 +70,50 @@ setMethod(
     )
     sapply(dirs, dir.create, recursive = TRUE, showWarnings = FALSE)
     
-    options_runtime <- substitute(
+    expr <- substitute(
       list(
-      rapp_global = rapp_global,
-      runtime_mode = "dev",
-      lib = .libPaths()[1]
-      ),
-      list(
-        rapp_global = file.path(Sys.getenv("HOME"), "rapp")
+        rapp_global = file.path(Sys.getenv("HOME"), "rapp"),
+        runtime_mode = "dev",
+        lib = .libPaths()[1]
       )
-    )
-#     write(deparse(options_runtime, control = NULL), file = "test.r")
-#     formatR::tidy_source(
-#       source = "test.r",
-#       brace.newline = TRUE,
-#       indent = 2,
-#       file = "test2.r",
-#       width.cutoff = 60
-#     )
-    
+    )    
     fpaths <- c(
       "rapp/options/options_runtime.r",
       "rapp/apps/test/options/options_runtime.r"
     ) 
-    idx <- which(!sapply(fpaths, file.exists))
+    if (overwrite) {
+      idx <- seq(along = fpaths)
+    } else {
+      idx <- which(!sapply(fpaths, file.exists))
+    }
     if (length(idx)) {
       sapply(fpaths[idx], function(ii) {
         write(
-          .formatOptionFile(x = options_runtime, name = "options_runtime"), 
+          rapp.core.rte::tidySource(input = expr, name = "options_runtime"), 
           file = ii
         )
       })
     }
-
-    options_app <- list(
-      ns = rapp.core.package::asPackage(x = ".")$package,
-      option_2 = "your option value here (can be any R object)",
-      option_3 = "your option value here (can be any R object)"
-    )
+    expr <- substitute(
+      list(
+        ns = rapp.core.package::asPackage(x = ".")$package,
+        option_2 = "your option value here (can be any R object)",
+        option_3 = "your option value here (can be any R object)"
+      )
+    ) 
     fpaths <- c(
       "rapp/options/options.r",
       "rapp/apps/test/options/options.r"
     ) 
-    idx <- which(!sapply(fpaths, file.exists))
+    if (overwrite) {
+      idx <- seq(along = fpaths)
+    } else {
+      idx <- which(!sapply(fpaths, file.exists))
+    }
     if (length(idx)) {
       sapply(fpaths[idx], function(ii) {
         write(
-          .formatOptionFile(x = options_app, name = "options"), 
+          rapp.core.rte::tidySource(input = expr, name = "options"), 
           file = ii
         )
       })