+++ inst/examples/ensureRuntimeEnvironment.r
@@ -1,7 +1,7 @@
 \dontrun{
 
 ## Expects certain components of the rapp framework to be in place!  
-ensureRappRuntimeEnvironment(
+ensureRuntimeEnvironment(
   rapp_global = "q:/home/rapp",
   runtime_mode = "dev"
 )