+++ inst/examples/getRappOption.r
@@ -4,7 +4,7 @@
 opts_old <- getOption(".rapp")
 
 ## Create example content //
-container <- ensureInitialRappOptions()
+container <- initializeRappOptions()
 cont_this <- getRappOption(id = ".rte")
 
 setRappOption(id = ".rte/test", value = TRUE)