+++ rapp/options/options_runtime.r
@@ -1,5 +1,5 @@
 options_runtime <- list(
-	rapp_global = "C:/Users/rappster_admin/Documents/rapp",
+	rapp_global = "q:/home/rapp",
 	runtime_mode = "dev",
 	lib = .libPaths()[1]
 )