+++ tests/testthat/test-resetRappOptions-1.r
@@ -4,7 +4,7 @@ test_that("resetRappOptions", {
   opts_old <- getOption(".rapp")
   
   expect_is(res <- resetRappOptions(), "environment")
-  expect_equivalent(res, ensureInitialRappOptions())
+  expect_equivalent(res, initializeRappOptions())
   
   on.exit(options(".rapp" = opts_old))